// src/interfaces/property/PropertyAsserts.ts
export default interface PropertyAsserts {
    id: number;
    propertyName: string;
    description: string;
    numberOfBuildings: number;
    numberOfFloors: number;
    numberOfUnitsRooms: number;
    buildingClassId: number;
    buildingStatusId: number;
    yearOfCompletion: string;
    yearRenovated: string;
    locationTypes: Array<number>;
    excessLand: boolean;
    greenRatings: Array<number>;
    propertyMarketId: number;
    propertySubMarketId: number;
    assetTypes: Array<number>;
    assetSubTypeId: number;
    landTenureId: number;
    leaseholdStartDate: string;
    leaseholdExpirationDate: string;
    leasehold: number;
    anchorTenants: string;
    buildingGradeId: number;
    landArea: number;
    landAreaUnitOfMeasureId: number;
    grossFloorArea: number;
    grossFloorAreaUnitOfMeasureId: number;
    sectors: Array<number>;
    netLeasableArea: string;
    netLeasableAreaUnitOfMeasureId: number;
    grossLeasableArea: number;
    grossLeasableAreaUnitOfMeasureId: number;
    carparkLots: string;
    majorTenants: string;
    passingRent: number;
    passingRentCurrencyId: number;
    passingRentUnitOfMeasureId: number;
    netOperatingIncome: number;
    capitalisationRate: number;
    ownershipTypeId: number;
    partialInterest: string;
    map: {
        long: number;
        lat: number;
        address: string;
    };
    imageUrl: string;
    isPortfolio: boolean;
    createdDate: string;
    updatedDate: string;
    createdById: string;
    status: number;
}